added tests with real world config files
