+++ packages/blunderbuss/test/fixtures/config/python-docs-samples-broken.yml
@@ -1,0 +1,164 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+assign_issues_by:
+- labels:
+  - 'api: appengine'
+  to: 'engelke'
+- labels:
+  - 'api: automl'
+  - 'api: dialogflow'
+  - 'api: language'
+  - 'api: speech'
+  - 'api: texttospeech'
+  - 'api: translate'
+  - 'api: videointelligence'
+  - 'api: vision'
+  to:
+  - sirtorry
+  - telpirion
+- labels:
+  - 'api: dataflow'
+  to:
+  - davidcavazos
+- labels:
+  - 'api: bigquery'
+  - 'api: datacatalog'
+  to:
+  - shollyman
+- labels:
+  - 'api: bigtable'
+  to:
+  - billyjacobson
+- labels:
+  - 'api: billingbudgets'
+  to:
+  - GoogleCloudPlatform/billing-samples-maintainers
+- labels:
+  - 'api: cloudsql'
+  to:
+  - GoogleCloudPlatform/cloud-sql-dpes
+- labels:
+  - 'api: containeranalysis'
+  to:
+  - donmccasland
+  - dinagraves
+  - kelsk
+- labels:
+  - 'api: datalabeling'
+  - 'api: iot'
+  to:
+  - sgreenberg
+- labels:
+  - 'api: dataproc'
+  to:
+  - bradmiro
+- labels:
+  - 'api: datastore'
+  - 'api: firestore'
+  to:
+  - GoogleCloudPlatform/firestore-dpe
+- labels:
+  - 'api: dns'
+  to:
+  - michaelawyu
+- labels:
+  - 'api: functions'
+  to:
+  - ace-n
+  - grant
+- labels:
+  - 'api: healthcare'
+  to:
+  - noerog
+- labels:
+  - 'api: logging'
+  - 'api: clouderrorreporting'
+  to:
+  - Daniel-Sanche
+- labels:
+  - 'api: ml'
+  to:
+  - alecglassford
+- labels:
+  - 'api: notebooks'
+  to:
+  - sirtorry
+- labels:
+  - 'api: cloudprofiler'
+  to:
+  - kalyanac
+- labels:
+  - 'api: pubsub'
+  to:
+  - anguillanneuf
+- labels:
+  - 'api: run'
+  to:
+  - averikitsch
+  - grant
+- labels:
+  - 'api: cloudscheduler'
+  - 'api: cloudtasks'
+  - 'api: tasks'
+  to:
+  - averikitsch
+- labels:
+  - 'api: spanner'
+  to:
+  - larkee
+- labels:
+  - 'api: storage'
+  to:
+  - GoogleCloudPlatform/storage-dpe
+- labels:
+  - 'api: cloudtrace'
+  - 'api: trace'
+  to:
+  - dukeng
+  - ziweizhao
+- labels:
+  - 'api: cloudiot'
+  to:
+  - fjanosi-google
+  - joeklemm
+- labels:
+  - 'api: datascienceonramp'
+  to:
+  - leahecole
+  - bradmiro
+
+assign_prs_by:
+- labels:
+  - 'api: cloudiot'
+  to:
+  - fjanosi-google
+  - joeklemm
+
+
+assign_issues:
+  - busunkim96
+  - leahecole
+  - engelke
+  - kurtisvg
+  - dinagraves
+  - dandhlee
+
+assign_prs:
+  - busunkim96
+  - leahecole
+  - engelke
+  - kurtisvg
+  - dinagraves
+  - dandhlee
